requirements.txt
scipy
numpy
pandas
seaborn
PyYAML
multiprocess
